Revert all changes made by Greg Woods today except the changes to contrib/log_accum.pl and contrib/.cvsignore.
This is for two reasons:

1) The changes do not separate out unrelated changes into separate
checkins and separate ChangeLog entries.  Just to list a few--the
changes regarding the makefile rule for configure, the changes
regarding the makefile rule for building the Makefile, the changes to
nuke options.h.in, adding RSH_DFLT, adding -R, adding CVS_SERVER_DFLT,
VERSION and PACKAGE in configure.  Each of these would need to be a
separate checkin and separate ChangeLog entry.

2) A number of the changes are controversial (as defined in HACKING).