Backout this patch:
	* rcs.c (findhighestmagicrev_proc): New function to find the
	highest magic rev with the required number of dots.
	(findnextmagicrev): New function to walk the symbolic tag list
	calling findhighestmagicrev_proc() to find the next unused magic
	branch revision.
	(RCS_magicrev): Use findnextmagicrev().
	(Patch suggested by "Kelly F. Hickel" <kfh@mqsoftware.com> based
	on advice from Larry Jones.)

As implemented, this function finds the highest rev across *all*
branches with the right number of dots, not just branches rooted at
the same revision. So, for example, if you have a number of branches
off 1.1 -- 1.1.0.2, 1.1.0.4, and 1.1.0.6, for example -- and then
create a new branch off of 1.2, you'll get 1.2.0.8 instead of 1.2.0.2
like you should. In fact, it checks *all* tags with the right number
of dots, not just magic branch tags, so plain revision tags will
pollute the branch number, too.